fix too many renders, doesn't render on geolocation change
