+++ next-run/public/index.html
@@ -23,6 +23,15 @@
     -->
     
     <title>Next Run</title>
+
+    <!-- <script async defer src="https://maps.googleapis.com/maps/api/js?v=weekly&key=AIzaSyDgrD2T5T3aXS3rG7hHrW5dtabt4gicJGk"
+    type="text/javascript"></script> -->
+
+    <!-- <script async defer
+    src="https://maps.googleapis.com/maps/api/js?v=weekly
+        &key=AIzaSyDgrD2T5T3aXS3rG7hHrW5dtabt4gicJGk&callback=initMap"> -->
+</script>
+
   </head>
   <body>
     <div id="root"></div>
